build for 2.7 on Windows

+++ .travis/deploy.sh
@@ -11,8 +11,11 @@ mkdir dist
 python setup.py sdist
 
 if [[ "$TRAVIS_OS_NAME" == "linux" && ${BUILD_LINUX_WHEELS} -eq 1 ]]; then
-	docker run --rm -v $(pwd):/io quay.io/pypa/manylinux1_x86_64 /io/.travis/build-linux-wheels.sh
-	linux32 docker run --rm -v $(pwd):/io quay.io/pypa/manylinux1_i686 /io/.travis/build-linux-wheels.sh
+    if [ "$TRAVIS_PYTHON_VERSION" == "3.5" ]
+    then
+        docker run --rm -v $(pwd):/io quay.io/pypa/manylinux1_x86_64 /io/.travis/build-linux-wheels.sh
+	    linux32 docker run --rm -v $(pwd):/io quay.io/pypa/manylinux1_i686 /io/.travis/build-linux-wheels.sh
+    fi
 	.travis/build_windows_wheels.sh
 else
 	# Only build wheels for the non experimental bundled version
